requirements.txt
absl-py==1.3.0
-e git+https://github.com/dyuliu/aer-paper.git@bb120f96d68eefeb9871033b636ac9a3a043b5e8#egg=aer
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
attrs==22.1.0
azure-cognitiveservices-anomalydetector==0.3.0
azure-common==1.1.28
azure-core==1.26.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
botocore==1.28.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
dask==2022.2.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distributed==2022.2.0
entrypoints==0.4
fastjsonschema==2.16.2
featuretools==0.22.0
fsspec==2022.10.0
gast==0.3.3
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.50.0
h5py==2.10.0
HeapDict==1.0.1
idna==3.4
imageio==2.22.2
importlib-metadata==5.0.0
importlib-resources==5.10.0
ipykernel==6.16.1
ipython==7.34.0
ipython-genutils==0.2.0
iso639==0.1.4
isodate==0.6.1
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.16.0
jupyter-server==1.21.0
jupyter_client==7.4.3
jupyter_core==4.11.2
jupyterlab-pygments==0.2.2
Keras==2.4.3
Keras-Preprocessing==1.1.2
langdetect==1.0.9
lightfm==1.16
llvmlite==0.34.0
locket==1.0.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
mlblocks==0.4.1
mlprimitives==0.3.2
mongoengine==0.24.2
msgpack==1.0.4
msrest==0.7.1
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.2
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.6.3
nltk==3.7
notebook==6.5.1
notebook_shim==0.2.0
numba==0.51.2
numpy==1.18.5
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
orion-ml @ git+https://git@github.com/sintel-dev/Orion.git@afbe23dee57e7f6e7090b89658e390997a1f42ac
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
partd==1.3.0
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
prometheus-client==0.15.0
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
pymongo==3.12.3
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-louvain==0.13
pyts==0.10.0
pytz==2022.5
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==24.0.1
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3fs==0.4.2
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
statsmodels==0.12.2
tabulate==0.8.10
tblib==1.7.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.3.4
tensorflow-estimator==2.3.0
termcolor==2.0.1
terminado==0.16.0
threadpoolctl==3.1.0
tifffile==2021.11.2
tinycss2==1.2.1
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.2.2
wrapt==1.14.1
xgboost==0.90
XlsxWriter==1.3.9
zict==2.2.0
zipp==3.10.0
